+++ http/src/main/scala/org/broadinstitute/dsde/workbench/leonardo/db/RuntimeServiceDbQueries.scala
@@ -139,4 +139,7 @@ object RuntimeServiceDbQueries {
       }.toList
     }
   }
+
+  def isDiskAttachedToRuntime(disk: PersistentDisk): DBIO[Boolean] =
+    ???
 }
